Update patch set 2
Patch Set 2:

I'm not sure what this buys you.  I think everywhere we print the client name, we print the cmd line if it's a user program of the name if it's not.  Is that not sufficient?  I'm not sure why it would be beneficial to store a string form of the pid when that can be retrieved from the task struct on demand.

Patch-set: 2
Reviewer: Gerrit User 1001118 <1001118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0