Update patch set 2
Patch Set 2:

I submitted this patch upstream, but it changed in implementation as the audit system gained a more generic system for feature implementation, and the KConfig was dropped. However, according to Eric Paris, the audit maintainer, this will be staged to Linus's tree for 3.11.

So this leaves us with two options, take this until this time, or back port the 3.11 audit system.

Patch-set: 2
Label: Verified=0