Update patch set 2
Patch Set 2:

Ack. Let me know how the attempt to write the test goes. Bear in mind that getting this change into android-3.4 will not by itself get this change into actual devices, because it needs to be cherry-picked into device kernel trees as well. If we have a test, it will be easy to tell if the change is in a particular device tree or not.

Patch-set: 2
Label: Code-Review=0