Update patch set 3
Patch Set 3:

Thanks Colin. We got strong negative feedback from Greg KH when we tried to submit the original patch to upstream kernel complaining about '#if defined' in .c files.
This is why I refactored the code to remove #ifdef from .c files. I am still not sure if this is acceptable to upstream kernel. 
If Patchset 3 is good from Google's end, we will run a validation cycle on it with on Fugu for the M upgrade and then do an +1. You can then merge the patch.

Patch-set: 3
Label: Verified=0