Update patch set 3
Patch Set 3:

Can we get this merged? it's already in the Android 3.10, 3.14, and 3.18 kernels as well as the upstream kernel. This does not need to be  CTS tested, it removes a spurious but harmless denial. It would be satisfactory if only new devices branching off the Android kernel had the patch.

Patch-set: 3
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0