Update patch set 3
Patch Set 3:

A CTS test is not possible because it requires reading the kernel log and it's not appropriate because this isn't a security bug so we would not enforce backporting. This exists in the upstream kernel, the other Android common kernels, and is running on most of our Nexus devices. What additional value would a unittest provide?

Patch-set: 3