Update patch set 2
Patch Set 2:

Yes FROMLIST: in title, and something like

(cherry picked from commit https://lkml.org/lkml/2015/12/21/337)

in the body of the message. Mix between putting this before the first line of commit body so it is prominent, or in the footer near their own additional signage and bug references, albeit breaking the footer defacto format. Some forgo the round brackets. So there is no real standard here. I follow the before-the-first-line-of-commit-body format that is prevalent in the Android user-space code commits.

Patch-set: 2