Update patch set 1
Patch Set 1:

> That is not what I'm suggesting. libbinder uses the same type for
 > NULL and non-NULL pointers, but does not tell the kernel about the
 > objects that contain NULL pointers.

So you're suggesting to write the meta-data in the data part of the parcel, but not the object list (when the object is NULL)? I thought there was value for the receiving process to know the kernel verified the meta-data object, but thinking about it more I guess there isn't any, in the case of nullptr.

Patch-set: 1