Update patch set 1
Patch Set 1:

> Do we want to debug them separately? If a binder call ends up in in
 > a hal that blocks waiting for a hwbinder call, it seems more useful
 > to have that information nearby.

It think it depends on what you're trying to debug. In a lot of cases, we're not interested in what the other domain is doing. In the cases that we are, it's pretty easy to just print output from both domains.

Patch-set: 1
Reviewer: Gerrit User 1013030 <1013030@85c56323-6fa9-3386-8a01-6480fb634889>