Update patch set 1
Patch Set 1:

> I'm not sure why it's not tagged stable.
 > 
 > fyi, I sent the patch to Jaegeuk about two weeks ago. Here's the
 > link.
 > https://sourceforge.net/p/linux-f2fs/mailman/message/35806928/

Ah, you didn't add a "cc: stable" marking, that's why.  Take a look at Documentation/stable_kernel_rules.txt for how to do this.

As this is now in Linus's tree, please email stable@vger the commit id of the patch in Linus's tree and ask the people there to merge it to the stable kernels, and cc: the f2fs maintainer as well.  I can take it from there...

Patch-set: 1
Reviewer: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>