Update patch set 1
Patch Set 1:

Please provide a bit more context here than the bug numbers, as people in AOSP generally do not have access to this data.

A simple, "filesystem accesses go a lot faster so we strongly recommend it being enabled" is fine, if that's the case, but you need to give us something...

Same for all kernel branches.

Patch-set: 1
CC: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>