Update patch set 1
Patch Set 1:

Sorry for the spam, turns out all these are unnecessary because CONFIG_AIO is on by default. I didn't realize that the config parsed comments so I assumed that lack of "CONFIG_AIO=y" meant that aio was off, when really "# CONFIG_AIO is not set" turns it off.

Patch-set: 1
CC: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>