ANDROID: binder: make sure accesses to proc/thread are safe
binder_thread and binder_proc may be accessed by other
threads when processing transaction. Therefore they
must be prevented from being freed while a transaction
is in progress that references them.

This is done by deferring the freeing of proc/thread
structures until binder_free_transaction() if part
of a transaction.

It is safe to free a binder_thread if its
transaction_stack is NULL.

It is safe to free a binder_proc if it has no
remaining threads and is not part of a
transaction (t->to_proc). To track this,
an atomic proc->txn_in_progress is added
that is incremented when building the
transaction and decremented when the transaction
is freed.

Change-Id: I0a00a0294c3e93aea8b3f141c6f18e77ad244078
Signed-off-by: Todd Kjos <tkjos@google.com>