sched/fair: compute dead-zone margin only if SchedTune is not in use
The dead-zone filtering is used to reject scheduling candidate where
the E variations due to moving a task from the prev_cpu to a candidate CPU
are below a certain margin, which is defined to be ~1.56% relative to the
maximum energy consumption computed for the current evaluation.

However, this margin is completely ignored in case SchedTune is in use, since
eenv->nrg_delta is updated by the __update_perf_energy_deltas() function.

Let's compute the dead zone only when SchedTune is not in use.

Change-Id: I4fbeeaf54fd486bd15606f8efd1c1c6be8f2c84b
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>