sched/fair: remove not necessary rcu_dereference
Both find_best_target() and energy_diff() need to walk the SDs and they
have their own rcu_dereference depending on the starting CPU.

It it not completely clear what is the role of the rcu_dereference in
select_energy_cpu_brute, which is also checked after find_best_target
has returned.

Let's remove it to have a better defined behavior:

- selecte_energy_cpu_brute defined the critical section star/end points
- find_best_target() returns target_cpu = -1 in case it fails to get a
  SD reference, thus returning prev_cpu from select_energy_cpu_brute()
- energy_diff() returns 0 in case if fails to get a SD reference, thus
  returning prev_cpu from select_energy_cpu_brute()

Thus, in general the select_energy_cpu_brute() function selects the
previous task CPU in case there are issue accessing the SD and/or EM
data.

Change-Id: I3a89127e4570f19cde84d94da061af0b9861463c
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>