sched/rt: add utilization tracking
schedutil governor relies on cfs_rq's util_avg to choose the OPP when cfs
tasks are running. When the CPU is overloaded by cfs and rt tasks, cfs tasks
are preempted by rt tasks and in this case util_avg reflects the remaining
capacity that is used by cfs tasks but not what cfs tasks want to use. In such
case, schedutil can select a lower OPP when cfs task runs whereas the CPU is
overloaded. In order to have a more accurate view of the utilization of the
CPU, we track the utilization that is used by RT tasks.
DL tasks are not taken into account as they have their own utilization
tracking mecanism.

We don't use rt_avg which doesn't have the same dynamic as PELT and which
can include IRQ time.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
[backported from list: http://marc.info/?l=linux-kernel&m=149786230529220]
Signed-off-by: Juri Lelli <juri.lelli@arm.com>
Change-Id: Icf36a67ea516b1ab21d9443a70e4b253f248550f
Signed-off-by: Chris Redpath <chris.redpath@arm.com>