Update patch set 1
Patch Set 1: Code-Review-2

I don't see a "why take this series" anywhere...

So, why do you feel we need this in AOSP?  What functionality that everyone can benefit from is missing that this will help enable?

Why shouldn't this just be in an SoC specific tree?  Why are we going to be willing to maintain this code now for 6+ years?

Patch-set: 1
Reviewer: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2