Update patch set 1
Patch Set 1: Code-Review+1

> I don't see a "why take this series" anywhere...
 > 
 > So, why do you feel we need this in AOSP?  What functionality that
 > everyone can benefit from is missing that this will help enable?
 > 
 > Why shouldn't this just be in an SoC specific tree?  Why are we
 > going to be willing to maintain this code now for 6+ years?

Hi Greg, as you probably know there are TrustZone on practically all ARM devices today. There are a few TrustZone implementations today leveraging this feature on the ARM chip. To name a few, QSEE, Huawei, Trustonic, Solacia, OP-TEE, Trusty etc. It is used in billions of mobile handsets, tablets etc already today.

In Android, TrustZone is used by Keymaster, Fingerprint reader, GateKeeper etc. So already today when you log into your mobile handset, your credentials will be stored using this functionality without you even notice or thinking about it.
https://source.android.com/security/authentication/

It's is also a requirement to use hardware backed secure environment to fulfill the Android Widevine Level 1 (meets current studio requirements). There are lots of other uses cases (keys for hard-disc encryption, device authentication, IMA etc), but the ones I mentioned here are the main ones for Android.

So, with the background and to answer the question: "why take this series". All current TrustZone/TEE implementation has been living out of tree until a month ago when we finally after submitting patches for two years to LKML managed to the generic TEE framework and also the OP-TEE driver upstream into Linux 4.12. With that we hope that this is the start of using functionality already upstream instead of having everyone in this area working with out of tree patches. The feedback from our members in Linaro using OP-TEE and other TEE's is that this is the right thing to do and they want to see this functionality officially in kernel. So, we see it as a natural step to merge this base functionality for TrustZone to Android also.

"Why shouldn't this just be in an SoC specific tree?"
It is not SoC specific, it is generic code for all platforms. We use the same kernel patches for +20 different platforms today (https://github.com/OP-TEE/optee_os#3-platforms-supported).

I hope this answer motivates why we should take this series.

Patch-set: 1
Reviewer: Gerrit User 1027838 <1027838@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1