Clear sk_dst_cache when applying per-socket XFRM policy.
The upstream patch in [1] skips XFRM lookups if a socket has a valid dst
cache. However, this cache is not cleared when applying policy to a
socket. The result is that new policies are ignored on sockets in
certain states.

This can be demonstrated like so,
1. Create UDPv6 socket.
2. connect() the socket.
3. Apply a per-socket XFRM policy.
4. send() data on the socket.

Packets will continue to be sent in the clear instead of matching an
xfrm or returning a no-match error.

Note: Creating normal XFRM policies should have a similar effect on
sk_dst_cache entries that match the policy, but that is not fixed in
this patch.

[1] commit 00bc0ef5880dc7b82f9c320dead4afaad48e47be
    "ipv6: Skip XFRM lookup if dst_entry in socket cache is valid"

Bug: 63589547
Test: This fixes tests in android-review.googlesource.com/c/418659
Change-Id: I2be67dba458f284a547aa713d3d6df1ccea30582
Signed-off-by: Jonathan Basseri <misterikkit@google.com>