Update patch set 5
Patch Set 5:

> My concern was this code assumes smp and not big little so the same optimization may not hold.

I think this code should generally work pretty well on big.LITTLE as it uses the group capacities.

> Just curious what's the outcome of that discussion with Morten ?

Oops, I think I dropped a sentence from my reply. I meant to say that we were thinking for a while and in the end decided that it always makes sense to use as many CPUs as possible, from a throughout POV.

Patch-set: 5