Update patch set 1
Patch Set 1:

> My plan is to remove the upstream Android kernel config fragments
 > as well, I can do that first if people would prefer. The config in
 > the mainline kernel really does not serve any purpose (other than
 > generating some confusion IMO) given that Android does not work on
 > tip. If Android does ever work on tip they can be added back.
 
So, I know I and a few others find having the fragment tied into the kernel tree as quite useful, especially with mainline (where despite it not being fully CTS compatible, it is still highly functional and very useful for testing).

Bummer I didn't see this last week as it would have been good to chat to understand the details motivating this better. Just to clarify, my understanding from your comment is basically: the definitive version of the configs are maintained in the test tool, and keeping the two in sync is a pain so your exporting it to its own shared project. Is that accurate?

Patch-set: 1
Reviewer: Gerrit User 1010110 <1010110@85c56323-6fa9-3386-8a01-6480fb634889>