Update patch set 1
Patch Set 1:

Spoke with John on IRC but for posterity, the evolution of this was that in O, the android-base.cfg kernel settings became required and enforced by VTS. Initially copies of the android-base.cfg from the three officially supported kernel versions were copied into VTS. This was bad enough as it meant keeping two sets of kernel configs in sync. Then the compatibility matrix code in system/libvintf started referring to the kernel configs in VTS as well, which is ugly in another way.

By centralizing the kernel configs in their own repo we're hoping to clean all this up. It should also help with the issue that there really is no true upstream for these configs right now. Commercialization tends to direct a lot of changes to the kernel configs to the ones in the android-common kernels, but the config fragments in Linus' tree get updates from a different set of people. There's a better chance IMO at keeping everything in sync if we keep them in one repository, where we can also conveniently update config fragments for all kernel versions in a single commit.

Patch-set: 1
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0