Update patch set 2
Patch Set 2:

This won't cause problems with other portions of the scheduler that use the enum flags as a boolean. The only issue it addresses is the one with the WALT accounting that caused the crash reported by Leo and others. You don't need this in products unless you have the cumulative_runnable_avg patch that caused the crash, which is what we're trying to fix here.

That's also why upstream didn't want the patch in 2015 - there's no need for the enum in the RT code in upstream, it can deal with on_rq as a boolean.

Patch-set: 2
Reviewer: Gerrit User 1093953 <1093953@85c56323-6fa9-3386-8a01-6480fb634889>