ANDROID: staging: android: Add fwdata driver
This driver reads certain Android-specific device properties (key-
value pairs) from platform firmware (e.g. ACPI tables), and
presents them to user space via a sysfs interface that mimics the
directory layout of /proc/device-tree/firmware/android/ as created
by the Device Tree (DT) driver.

The focus of this patch is on device properties related to Android
fstab configuration, which Android Oreo reads during the first
stage of init so as to mount /system and /vendor partitions early.
Platforms that use ACPI instead of DT, especially those based on
Intel architecture (such as Ranchu x86/x86_64 of Android Emulator),
should enable this driver and ask /init to look for first-stage
mount configuration in the new sysfs directory:

 /sys/bus/platform/devices/ANDR0001:00/properties/android/

For example, Android Emulator makes the following changes to its
Ranchu x86/x86_64 virtual board that work in tandem with the fwdata
driver:

 https://r.android.com/442472
 https://r.android.com/443432

Device properties are read from firmware using the unified device
property API:

 https://lwn.net/Articles/612062/

In the current implementation, this driver creates sysfs nodes
statically, and hard-codes all property keys. A future patch should
implement a smarter mechanism which populates the sysfs directory
dynamically, so that no change to this driver is necessary in case
we want to support more properties.

This driver was originally submitted to kernel/goldfish.git on AOSP
as goldfish_fwdata:

 https://r.android.com/442393

which prevented it from being used outside Android Emulator. This
patch ports it to kernel/common.git as android_fwdata, and makes a
few changes to the original version:

 1) Remove the dependency on the Ranchu virtual platform
    (CONFIG_GOLDFISH).
 2) Move the source files to drivers/staging/android/.
 3) Fix style issues caught by checkpatch.pl.

Test: Boot patched common 4.9 kernel (x86_64_ranchu_defconfig +
      CONFIG_ANDROID_FWDATA) in Android Emulator, verify the new
      sysfs-based Android DT directory is created, together with
      files that represent first-stage mount configuration for
      /system and /vendor partitions.

Change-Id: I8b6e4e933a3fa298189da4ba055921f3c294ebcc
Signed-off-by: Yu Ning <yu.ning@intel.com>