Update patch set 7
Patch Set 7: Code-Review+1

I'm still happy with this and as said in my previous comment a while ago, I think this is the right thing to do (and Linaro members working with TEE's are supporting this too).

It'd be good to hear anything from the Google maintainers to know if they agree or not, just to know if we should a) drop this, b) re-work the patch set or c) just merge it as it is.

Patch-set: 7
Reviewer: Gerrit User 1027838 <1027838@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1