Update patch set 1
Patch Set 1:

Change LGTM, thanks for backporting this!

In order for this to actually get into the field we'll need to test for it somehow. What is the actual impact of this change? Can we engineer a net_test or VTS test for it?

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>