PM / OPP: list_del_rcu should be used in function _remove_list_dev
list_del_rcu should be used to replace list_del in function _remove_list_dev,
since the opp is a rcu protected pointer.

Test: For example, on an ARM big.Little platform, the opp_table of big cluster
will be removed when big cluster is removed, which should be implemented in
the cpufreq driver.
Then A stress test that the big cluster is plugged in/out, can test the patch.

Change-Id: Ie7a5097b1149aa46177403b91798bfa9553b6442
Signed-off-by: Vincent Wang <vincent.wang@spreadtrum.com>