Update patch set 1
Patch Set 1: Code-Review-1

(11 comments)

I've found a bunch of tiny things, if this is really going to be a driver to merge.

But first, why not submit it upstream and then backport it after it gets merged?

Also, why is this needed at all?  Can't you get this information already today from sysfs in the firmware tree that ACPI exports?  Why does this need to be specific for Android, nothing in the data being exported seems specific to me.

Patch-set: 1
Reviewer: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1