BACKPORT: arm64: cpu_errata: Add Kryo to Falkor 1003 errata
The Kryo CPUs are also affected by the Falkor 1003 errata, so
we need to do the same workaround on Kryo CPUs. The MIDR is
slightly more complicated here, where the PART number is not
always the same when looking at all the bits from 15 to 4. Drop
the lower 8 bits and just look at the top 4 to see if it's '2'
and then consider those as Kryo CPUs. This covers all the
combinations without having to list them all out.

Fixes: 38fd94b0275c ("arm64: Work around Falkor erratum 1003")
Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
(cherry picked from commit bb48711800e6d7aedbf4dfd3367e0ab1270a6bea)

Change-Id: I740eaf8c6b3842e000e3f42500391534b3b4a63a
[ghackmann@google.com: remove bits related to errata scope, which
 doesn't exist in 4.4]
Signed-off-by: Greg Hackmann <ghackmann@google.com>