Update patch set 1
Patch Set 1:

(11 comments)

> But first, why not submit it upstream and then backport it after it gets merged?

Jin Qian asked me the same question before. I don't really like the way I implemented this driver (e.g. hard-coding all property keys is ugly), so I'm not confident the upstream would accept such a patch. Besides, I think there's an alternative solution that is probably more upstreamable - see below.

> Also, why is this needed at all?  Can't you get this information already today from sysfs in the firmware tree that ACPI exports?

That's what I thought in the beginning, but unfortunately, the ACPI driver doesn't export everything to sysfs like the DT driver does. ACPI allows device properties to be encoded in the _DSD (Device Specific Data) section of a Device node, but the ACPI driver doesn't create sysfs entries for anything in _DSD. To parse the properties stored in _DSD, one has to create a driver that matches the ACPI Device node and calls the Unified Device Property API:

https://lwn.net/Articles/612062/

Apparently, the ACPI driver assumes that device properties only need to be read by the corresponding device driver and are never to be exported to user space. However, Android expects otherwise.

I discussed the issue with a few colleagues at Intel before writing this driver, and we thought the best solution would be to modify the ACPI driver and make it export everything in _DSD to sysfs in DT style. The Google Android Emulator team (Bo Hu) preferred creating a new driver at the time (which is understandable in the circumstances), so that's what I ended up doing. But if we did it the other way and did it right (which could be difficult), we would probably have better chances of upstreaming (if the ACPI maintainers don't reject it as a crazy idea).

> Why does this need to be specific for Android, nothing in the data being exported seems specific to me.

Right. This driver starts by creating properties/android/, but it's possible to create properties/xyz/ as well. The thing is only Android expects device properties to be exported to user space in DT style, so I hesitated to make the driver more generic. I also wonder how I should name the driver and where to put it if I move it out of drivers/staging/android/.

Patch-set: 1
Reviewer: Gerrit User 1057369 <1057369@85c56323-6fa9-3386-8a01-6480fb634889>