Update patch set 1
Patch Set 1:

> Apparently, the ACPI driver assumes that device properties only need to be read by the corresponding device driver and are never to be exported to user space. However, Android expects otherwise.

There's no reason why we can not just add this to the ACPI core to export everything, that would benefit everyone instead of just one single platform :)

> I discussed the issue with a few colleagues at Intel before writing this driver, and we thought the best solution would be to modify the ACPI driver and make it export everything in _DSD to sysfs in DT style. The Google Android Emulator team (Bo Hu) preferred creating a new driver at the time (which is understandable in the circumstances), so that's what I ended up doing. But if we did it the other way and did it right (which could be difficult), we would probably have better chances of upstreaming (if the ACPI maintainers don't reject it as a crazy idea).

Please work with upstream to solve this generically, it should be trivial to take your existing code and do so.

> > Why does this need to be specific for Android, nothing in the data being exported seems specific to me.
> 
> Right. This driver starts by creating properties/android/, but it's possible to create properties/xyz/ as well. The thing is only Android expects device properties to be exported to user space in DT style, so I hesitated to make the driver more generic. I also wonder how I should name the driver and where to put it if I move it out of drivers/staging/android/.

Because of that reason, it's good to do this generically for all of ACPI.  Feel free to cc: me on any upstream patches if you want me to get involved in the review there.

Patch-set: 1
Reviewer: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>