Update patch set 1
Patch Set 1:

(3 comments)

> There's no reason why we can not just add this to the ACPI core to export everything, that would benefit everyone instead of just one single platform :)

It's great to have your support on this.

> Please work with upstream to solve this generically, it should be trivial to take your existing code and do so.
> Because of that reason, it's good to do this generically for all of ACPI.  Feel free to cc: me on any upstream patches if you want me to get involved in the review there.

Thanks a lot for taking your time to review this premature driver, as well as the advice and encouragement.

I also want to implement a proper, upstreamable solution, and I understand that I need to prepare another patch for the ACPI core driver and send it upstream for feedback. But I wonder how much of the current code can be reused in the new patch, which is supposed to be generic, so it can't hard-code any Android-specific device property. The current code, though, is almost entirely based on prior knowledge about the properties and the static creation of all the sysfs entries. Surely there's a big gap?

Since you are the expert on sysfs, I hope to get some tips (especially on the dynamic creation of sysfs entries) from you here before I write the new patch.

Patch-set: 1
Reviewer: Gerrit User 1057369 <1057369@85c56323-6fa9-3386-8a01-6480fb634889>