Update patch set 2
Patch Set 2: Code-Review-1

Why isn't this patch upstream first?  Why add a new kernel feature to the common tree that we have to maintain for the next decade or so, when it's easier to push it upstream and then we don't have to worry about it anymore?

As it is, adding new stuff is a pain (example, you missed the 4.14 branch that is being created...)

Patch-set: 2
Reviewer: Gerrit User 1125653 <1125653@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1