DO NOT SUBMIT: symbol namespaces example.
This patch shows a potential direction for introducing
symbol namespaces in the kernel.

The EXPORT_SYMBOL_NS() macro allows a symbol to be exported
to a specific namespace only. Entries exported this way are
stored in a new section in the kernel and module binaries,
called __ksymtab_ns. Having a single section for all
namespaces makes introducing new namespaces a lot easier,
at the cost of some asm() magic to encode the namespace
in the symbol itself.

The IMPORT_NS() macro allows a module to use symbols from
a particular namespace. These "allows" are stored in a
special __knsimports section.

This patch includes modifications to modpost to extract
and store the namespace with a symbol, and to extract the
allowed namespaces as well. This allows modpost to emit
warnings when a symbol is used that is part of a namespace
that was not imported by the module. (Obviously, we can
change the behavior to make this an error).

To show how this works, this patch modifies the
usb_poison_urb symbol to be a part of the USB namespace,
and modifies the modules using it to import the namespace
accordingly. Removing one of these IMPORT_NS statements
would lead to a warning like the following:

WARNING: module usbserial uses symbol usb_poison_urb
from namespace USB, but does not import it.

This is by no means a finished patchset, but I think it
provides a useful framework for continuing the discussion
and agreeing on a direction.

Known limitations/quirks:
* No EXPORT_SYMBOL_NS_GPL etc. yet
* IMPORT_NS() works on module granularity, not on source-
  file granularity. That means if a module is built from
  foo.c and bar.c, and both foo.c and bar.c use a symbol
  that is part of a namespace, strictly only one of them
  would need to import it. Perhaps it would be better to
  name it MODULE_IMPORT_NS(), so it can be placed with
  the other MODULE_ macros.
* Didn't change the module loaded yet, but it would need
  to be changed to deal with symbol namespaces at run-time
  (this shouldn't be hard).

Change-Id: Ic29e1c4e9a66bd35741a9cede6f0cb630b65eed8
Signed-off-by: Martijn Coenen <maco@android.com>