Update patch set 1
Patch Set 1:

> But doesn't this patch makes sense anyway? It isn't changing the current behaviour buy allows people to play easily with half-life period and not to re-invent. What's wrong in getting this merged to main android 4.9 itself ?

That is because the development model for EAS changed recently (Chris sent an email about that on the eas-dev mailing list in November IIRC). New features should go in android-4.9-eas-dev, which should be merged into android-4.9 twice a year. The reason behind this change is to be better aligned with the schedules of partners. I hope that makes sense :)

Patch-set: 1
Reviewer: Gerrit User 1200908 <1200908@85c56323-6fa9-3386-8a01-6480fb634889>