Update patch set 2
Patch Set 2:

Please be careful if we're really going to take these changes for all these different filesystems.  There is a follow-on fix for tmpfs and nfs that was missed (FYI, it is also missing from chromeos-4.4):

e8ecde25f5e0 ("Make sure that highmem pages are not added to symlink page cache")

Also page_readlink() is broken, since it calls kunmap() an a page that wasn't kmap()'ed.  (page_readlink() was no longer present when this patch was made upstream.)

Did you consider just doing the changes for ext4 and f2fs only?  If I understand correctly, the main reason Al Viro made this change is so that symlinks can be followed without dropping out of RCU mode.  But, that isn't implemented in older kernels.  So in old kernels it's really only needed for ext4 and f2fs, for following encrypted links.

(There was also a concern that with the old code, a malicious application could use up all available kmap slots, but that would be a denial of service only.)

Patch-set: 2
CC: Gerrit User 1137063 <1137063@85c56323-6fa9-3386-8a01-6480fb634889>