Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> Shouldn't something like this, that adds a new user/kernel api, go upstream first?  Do you really want to maintain this new API for the next 20+ years?  :)

This was discussed for the first implementation of this (wahoo/marlin kernels: https://partner-android-review.googlesource.com/c/kernel/private/msm-google/+/879133) and we felt this feature stands pretty much no chance of acceptance upstream since the way android uses UIDs is unusual so this model wouldn't be interesting outside the android community.

Looking at this series, what do you think? I totally agree that it would be better if it were upstream.

Accurate battery stats is an important enough feature that we'd be willing to support it in common for however long.

Patch-set: 5
Reviewer: Gerrit User 1069543 <1069543@85c56323-6fa9-3386-8a01-6480fb634889>